requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-genai==1.0.0
pinecone-client==5.0.1
pydantic==2.9.2